# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис безопасного хранения паролей, карт и файлов.",
        "title": "GophKeeper API",
        "termsOfService": "Free use",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ElfAstAhe/goph-keeper",
            "email": "elf.ast.ahe@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/users/keys": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Генерирует новую пару RSA ключей (приватный перешифровывается мастер-паролем) и возвращает новый публичный ключ",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Обновление мастер-ключей пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ElfAstAhe_goph-keeper_api_rest_dto.ChangeKeysResultDto"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_ElfAstAhe_goph-keeper_api_rest_dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ElfAstAhe_goph-keeper_api_rest_dto.ErrorDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_ElfAstAhe_goph-keeper_api_rest_dto.ChangeKeysResultDto": {
            "type": "object",
            "properties": {
                "public_key": {
                    "type": "string"
                }
            }
        },
        "github_com_ElfAstAhe_goph-keeper_api_rest_dto.ErrorDto": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "details": {
                    "type": "string",
                    "example": "field 'e-mail' is required"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid input data"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer \u003cJWT_TOKEN\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}